chore: never evict fee token asset metadata from cache (#1494)
`wallet_getCapabilities` will try and fetch the metadata of all its fee
tokens. If they're not in the cache because they got evicted, they'll be
fetched onchain. This can be problematic if any of the provider is
having issues

This ensures that they are always in the cache without ever getting
evicted.