chore: handle problematic endpoints (#1497)
We have been having some issues with problematic endpoints resulting in
the relay seemingly being stuck.

* adds a rpc request timeout config per chain
* `wallet_getAssets` now returns what it can and only logs errors. This
method is also used internally (eg. cross chain balance
check/discovery). So, if a chain endpoint is misbehaviing and a client
tries to query it, it will respond with no assets, even if that might
not be true.
* `wallet_getCapabilities` now returns what it can and only logs chain
errors. Shouldn't happen though with
https://github.com/ithacaxyz/relay/pull/1494